Refactor updateDraftContractRates handler into updateDraftContract.
