+++ services/app-api/src/domain-models/healthPlanPackage.ts
@@ -126,7 +126,9 @@ function convertContractRateRevisionToHealthPlanRevision(
         rateProgramIDs: formData.rateProgramIDs,
         rateCertificationName: formData.rateCertificationName,
         actuaryContacts: formData.certifyingActuaryContacts?.length
-            ? formData.certifyingActuaryContacts
+            ? formData.certifyingActuaryContacts.concat(
+                  formData.addtlActuaryContacts ?? []
+              )
             : [],
         // From the TODO in convertContractRevisionToHealthPlanRevision, these can just be set as whatever is in the
         // database. The frontend does not read this values.
@@ -150,8 +152,6 @@ function convertContractRevisionToHealthPlanRevision(
     //  the rate revision level. Since at update contract we are setting both fields in each rate using what the values
     //  on the contract level, when we pull data out from our new DB model, we need to do the inverse by using, the
     //  the first rates values.
-    //  This will need to be updated and fixed when we figure out shared rates, because shared rates are not
-    //  guaranteed to have the same values.
     const addtlActuaryCommunicationPreference =
         contract.draftRevision?.rateRevisions[0]?.formData
             ?.actuaryCommunicationPreference
