+++ services/app-api/src/postgres/contractAndRates/index.ts
@@ -1,11 +1,9 @@
 export type { InsertContractArgsType } from './insertContract'
 export type { UpdateContractArgsType } from './updateDraftContract'
 export type { ContractOrErrorArrayType } from './findAllContractsWithHistoryByState'
-export type { UpdateDraftContractRatesType } from './updateDraftContractRates'
 
 export { insertDraftContract } from './insertContract'
 export { findContractWithHistory } from './findContractWithHistory'
 export { updateDraftContract } from './updateDraftContract'
 export { findAllContractsWithHistoryByState } from './findAllContractsWithHistoryByState'
 export { findAllContractsWithHistoryBySubmitInfo } from './findAllContractsWithHistoryBySubmitInfo'
-export { updateDraftContractRates } from './updateDraftContractRates'
